// src/pages/applicationOrchestration/pageCanvas/components/addComponentNode/types.ts
export interface CommponentBeansItem {
    belongModule?: string;
    businessId?: string;
    componentCategory: string;
    componentDesc?: string;
    componentLevel?: string;
    componentName: string;
    componentStatus?: string;
    createStaffId?: string;
    createTime?: string;
    dataType?: string;
    id: string;
    provId: string;
    serviceLink: string;
    parentId: string;
    serviceTypeId?: string;
    updateStaffId?: string;
    updateTime?: string;
    componentPicture?: string;
}